Releasing `0.0.0dev12`
This change includes:

- `gitbolt` now fully lightweight.
- All operations are designed to be totally lazy. This allows for interacting with underlying system when absolutely
necessary.
- `git` typical operations such as `--html-path`, `--exec-path`, etc. previously interacted with the underlying
system on ctor call since these were all kept as properties and properties needed filling at the creation time.
This meant that `git` objects were heavyweight and cloning bogged down the system. Now properties are reduced to
method calls to make them lightweight and lazy.
- A step is being taken to provide users with programmatic returns of methods. Previously all methods simply returned
strings and preferred for the users to parse those strings, now the design is changed to make sense of the returned
rosetta strings and provide users with something that can be used and manipulated programmatically.
- Fixed bug [Vaastav-Technologies/3-check-not-respected-in-uncheckedsubcmd](https://github.com/Vaastav-Technologies/py-gitbolt/issues/3)
- some failing tests to be fixed on Mac OS.